[fx][quant] Add `intrinsic.quantized.LinearReLU` to the fx2trt converters
The fx2trt handles the fused operators by replacing them with the TRT layer followed by the TRT activation.
This adds the quantized LinearReLU conversion routine to the fx2trt.
This is required because the `torch.ao.quantization.convert_fx` replaces the Linear + ReLU with a fused LinearReLU operator that fx2trt should be aware of.

ghstack-source-id: 28dc0c2ce283c584ebed8e84a209d8d9bfcfa3b8
Pull Request resolved: https://github.com/pytorch/pytorch/pull/69573