Refactor convolution_backward's cudnn cases
Changed the Cudnn and CudnnTranspose cases to only make the input
contiguous when it is needed for the grad_weight computation.

Reading the implementation of cudnn_convolution_transpose_backward and
cudnn_convolution_backward give me confidence that `input` isn't used
for the grad_weight computation. However, the memory format logic is so
convoluted that I'm not 100% sure this correct. All the tests though
and on request I can directly pass `backend_memory_format` to
{cudnn_convolution_backward, cudnn_convolution_transpose_backward}.

Test Plan:
- pytest test/test_nn.py -v -k "conv"

ghstack-source-id: 63a54a1ce892714dec3b2e8de309cc6687a2627e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/71491