Relax dtype restrictions on torch.Tensor
Previously, torch.Tensor was treated as if it were torch.FloatTensor
(where Float is whatever the default dtype was).  This is not good
behavior for tensor subclasses, which inherit from torch.Tensor and
will want to super() call into it and will only notice later that
only float works as a dtype.  So in this PR I relax the behavior
for this case to make the torch.Tensor constructor more useful for
subclasses.

Signed-off-by: Edward Z. Yang <ezyangfb.com>

ghstack-source-id: de9c75420c1129694071b783fb1ae4f7f1adae90
Pull Request resolved: https://github.com/pytorch/pytorch/pull/73850