Disable __torch_function__ wrapping if __torch_dispatch__ is defined
This is never what you want and has led us to have to spray
`__torch_function__ = _disabled_torch_function_impl` everywhere.
Fortunately the fix is simple.

You might still want to overwrite `__torch_function__` for performance
reasons, as you can bypass having to run the Python code associated with
the default `__torch_function__`.  It might be a good performance
optimization to bypass this by default if we detect you have the
default `__torch_function__` and a `__torch_dispatch__`, but that is
left for future work.

Signed-off-by: Edward Z. Yang <ezyangfb.com>

ghstack-source-id: ab8e7d15d1e8498e36fbb73ca69729ef32e81f08
Pull Request resolved: https://github.com/pytorch/pytorch/pull/73942