[c10d] Use c10::Synchronized in c10d::SequenceNum
The new `c10::Synchronized` class helps enforce that locks are properly
held when touching guarded fields. Use it instead of a raw mutex.

Also add an apparently-missing TORCH_CHECK in get(). The header comment
says that it will throw if the number is not set, but get() didn't do
the check that the other methods did.

ghstack-source-id: 46918e266cfae51705c116af403c3b8b2d7b7286
Pull Request resolved: https://github.com/pytorch/pytorch/pull/74049