Remove structured kernels special casing in the codegen.
This PR removes the special casing introduced in #73350.

**What was this special casing for?**
It was introduced so that structured kernels could take advantage of `IList` from the
dispatcher. Basically, this special casing associated, for example, `Tensor[]` with
`const ITensorList &`, instead of the previous `TensorList`. The other kernels remained
using `TensorList`.

**Why is it being removed?**
After #72414 and #72414, this special casing is no longer needed. Those 2 PRs modifies the
signatures of every kernel that has `Tensor[]` or `Tensor?[]` as arguments. Replacing them
by `ITensorList` and `IOptTensorRefList`, respectively. From that PR onwards, every kernel
is using the new API, so there's no need to special case it only for structured kernels.

ghstack-source-id: a4710701d9ef0aac7b900cb3413aa8167d1cd634
Pull Request resolved: https://github.com/pytorch/pytorch/pull/73641