[quant][core] Add more fields for QConfig to address current use cases
Summary:
Current qconfig doesn't distinguish between input and output activation observer and use the same observer for input and output,
this means it does not support the case when input is quantized and output is float.
Also it doesn't support bias observer, this is OK for native backend (fbgemm/qnnpack) but may be needed
for custom backends.

This PR just added the qconfig, we'll add support in fx graph mode quantization in a future PR

Test Plan:
python test/test_quantization.py

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 7330a34e1957a0dfc573467e8e46a0fa8d229c26
Pull Request resolved: https://github.com/pytorch/pytorch/pull/75006