[SR] Make sub-blocks properly extend tensor lifetimes
Pull Request resolved: https://github.com/pytorch/pytorch/pull/75379

Another day, another liveness analysis bug.

This one is quite subtle and hard to trigger. It can only happen when the last use of a managed tensor (including all aliases) is buried inside a node with sub-blocks like `prim::If`. If this is the case, the lifetime of that tensor needs to be extended to the index of node with sub-blocks.

I've modified the first pass of the liveness analysis (the one that doesn't consider aliases) to consider nodes referenced from outside the sub-block's scope. So in this graph

```
b = some_op() # (node 1)
c = some_op() # (node 2)
if condition: # (node 3)
    x = another_op(b)
else:
    x = some_op()
ret = c.clone() # (node 4)
return ret
```

The lifetime of `b` is `[1, 3]` instead of `[1, 1]`. The memory planner correctly recognizes `b` as overlapping with `c`, and we return the correct result.

Nested sub-blocks are considered too; we would get the same result in the graph above if we instead had:
```
if condition1: # (node 3)
    if condition2:
        x = another_op(b)
```

There's another case we have to handle - nodes with sub-blocks may also extend lifetimes by producing a *new* alias of an outer scope value, as in the following graph:

```
def forward(self, a, cond: bool):
    b = torch.abs(a)
    if cond:
        d = b.view(a.size())
    else:
        d = b * b
        d = d * d # extra op to avoid ifThenElse optimization
    e = torch.argmin(a) # b is still alive, don't overwrite it!
    return d.clone(), e.clone()
```

The fix is simple. When we are traversing the graph backwards to extend lifetimes through aliases, for each node `n`, consider the set `s := {inputs(n)} union {valuesFromOuterScope(n)}`. A tensor `s_i \in s` extends the lifetime of tensor `x_j \in outputs(n)` if `mayAlias(s_i, x_j)`.

ghstack-source-id: 153261714

Differential Revision: [D35448776](https://our.internmc.facebook.com/intern/diff/D35448776/)