[quant][refactor] Remove the base class from __all__
In general, if we are expecting the users to use the base class,
such as `_ConvNd`, we should rename it to something like
`BaseConv`. However, because this base class is only used inside of the
AO packages, there is no need to expose it to the users.

Test Plan:

```
python test/test_quantization.py
python test/test_module_init.py
```

ghstack-source-id: 02f7241c6e5a73bbe60c7ca453d2afad24c6cd01
Pull Request resolved: https://github.com/pytorch/pytorch/pull/77344