[quant][core] Added support for quantizing a constant value and broadcasting the assignment to an entire tensor
Summary:
Support in QuantizedOpKernels.cpp exists for quantizing a
fp tensor and copying the quantized tensor to the output (quantized)
tensor. This PR adds analogous support for quantizing a scalar input and
broadcasting the quantized value to the entire tensor.

This PR is intended to be used for fill_quantized, where, currently, a
tensor filled with the same quantized value needs to be created in order
to be assigned to an output quantized tensor, resulting in an
inefficient fill operation.

Test plan:
See the subsequent PR as this PR does not make changes that would affect
front end code.

ghstack-source-id: 2bb69b58c0e8aed7338f19682aae55b1ce8f49f1
Pull Request resolved: https://github.com/pytorch/pytorch/pull/78562