Make slogdet, linalg.sloget and logdet support metatensors
This PR also adds complex support for logdet, and makes all these
functions support out= and be composite depending on one function. We
also extend the support of `logdet` to complex numbers and improve the
docs of all these functions.

We also use `linalg_lu_factor_ex` in these functions, so we remove the
synchronisation present before.

ghstack-source-id: 5808b6e73e641f05df6b76baff01e79da65682dc
Pull Request resolved: https://github.com/pytorch/pytorch/pull/79742