[ao][sparsity] Bug fix: data not correctly attached to the sparsifier
Issue:
Previously, the data was not "attached" to the data sparsifier. Meaning
the data sparsifier created a copy of the actual data inside it's container. So,
when the data was modified outside of the sparsifier, the changes was not reflected
in the sparsifier.

Fix:
Use register_buffer() instead of nn.Parameter(..) to store the data inside the container.
Also, added a unit-test to reference this issue.

Test Plan:
```python test/test_ao_sparsity.py TestBaseDataSparsifier```
```python test/test_ao_sparsity.py TestNormDataSparsifiers```

ghstack-source-id: 4c0531a8f22867091a2e67bd115b097c02af68e1
Pull Request resolved: https://github.com/pytorch/pytorch/pull/80394